chore: move plugin webpack config
